Resolved some bugs in remove user feature
NoW any user who is not a librarian and hasn't signed up can't perform delete operation. Also if the user who is being removed has borrowed books from the library, s/he won't be removed

+++ src/main/java/edu/achs/servlets/DeleteBanReturnOperations.java
@@ -7,6 +7,7 @@ package edu.achs.servlets;
 
 import edu.achs.daoImpl.BookDaoImpl;
 import edu.achs.daoImpl.UserDaoImpl;
+import edu.achs.entities.Users;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
@@ -26,28 +27,44 @@ public class DeleteBanReturnOperations extends HttpServlet {
 
     protected void processRequest(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        if (request.getRequestURI().contains("/deleteBook")) {
-            String bookId = request.getParameter("bookId");
-            if (new BookDaoImpl().isBookBorrowed(bookId)) {
-                request.setAttribute("errorMsg", "The book is borrowed by member(s).");
+
+        if (request.getSession().getAttribute("currentUser") != null) {
+
+            Users user = (Users) request.getSession().getAttribute("currentUser");
+
+            if (user.getUserType().equals("Librarian")) {
+                if (request.getRequestURI().contains("/deleteBook")) {
+                    String bookId = request.getParameter("bookId");
+                    if (new BookDaoImpl().isBookBorrowed(bookId)) {
+                        request.setAttribute("errorMsg", "The book is borrowed by member(s).");
+                    } else {
+                        new BookDaoImpl().deleteBook(bookId);
+                    }
+                    response.sendRedirect(request.getContextPath() + "/dashboard/books/all");
+
+                } else if (request.getRequestURI().contains("/user/ban")) {
+                    int userId = Integer.parseInt(request.getParameter("userId"));
+                    if (!new BookDaoImpl().isBookBorrowedByUser(userId)) {
+                        String profileImgName = new UserDaoImpl().getProfileImgName(userId);
+
+                        new UserDaoImpl().banUser(userId);
+
+                        if (!profileImgName.contains("_Default_pp.png")) {
+
+                            Files.deleteIfExists(Paths.get(request.getRealPath("Images") + File.separator + "ProfilePictures"
+                                    + File.separator + profileImgName));
+                        }
+                        request.setAttribute("successMsg", "User Removed Successfully!!");
+                    } else {
+                        request.setAttribute("errorMsg", "Could Not remove. User hasn't returned books!!");
+                    }
+                    response.sendRedirect(request.getContextPath() + "/dashboard/members/all");
+                }
             } else {
-                new BookDaoImpl().deleteBook(bookId);
+                request.setAttribute("errorMsg", "Invalid request");
+                response.sendRedirect(request.getContextPath() + "/home");
             }
-            response.sendRedirect(request.getContextPath() + "/dashboard/books/all");
-
-        } else if (request.getRequestURI().contains("/user/ban")) {
-            int userId = Integer.parseInt(request.getParameter("userId"));
-            String profileImgName = new UserDaoImpl().getProfileImgName(userId);
-
-            if (!profileImgName.contains("_Default_pp.png")) {
-                new UserDaoImpl().banUser(userId);
-                Files.deleteIfExists(Paths.get(request.getRealPath("Images") + File.separator + "ProfilePictures"
-                        + File.separator + profileImgName));
-            }
-            request.setAttribute("successMsg", "User Removed Successfully!!");
-            response.sendRedirect(request.getContextPath() + "/dashboard/members/all");
         }
-
     }
 
     // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
